requirements.txt
blinker==1.4
Flask==1.0.2
delegator.py==0.0.1
